requirements.txt
Django
djangorestframework
keytotext
numpy
setuptools
torch
django-cors-headers
sentencepiece
transformers
jinja2


